Alexander 6.0 Aligned with Sep 13, 2025 Revert "Adjustment of the aspiration window after fail high/low."
Various improvements and bug fixing

+++ src/uci.h
@@ -34,7 +34,7 @@ namespace Alexander {
 class Position;
 class Move;
 class Score;
-enum Square : int;
+enum Square : int8_t;
 using Value = int;
 
 class UCIEngine {
